+++ src/Coypu/ElementScope.cs
@@ -4,12 +4,16 @@ using Coypu.Queries;
 
 namespace Coypu
 {
+    /// <summary>
+    /// 
+    /// </summary>
     public interface ElementScope : Scope, Element
     {
         bool Exists(Options options = null);
         bool Missing(Options options = null);
         ElementScope Click(Options options = null);
-        Scope Hover(Options options = null);
+        ElementScope Hover(Options options = null);
+        ElementScope SendKeys(string keys, Options options = null);
     }
 
     public class DeferredElementScope : DriverScope, ElementScope
@@ -59,26 +63,57 @@ namespace Coypu
             get { return Now()[attributeName]; }
         }
 
-        public ElementScope Click(Options options = null)
+        public virtual ElementScope Click(Options options = null)
         {
-            RetryUntilTimeout(new Click(this, driver, SetOptions(options)));
+            ClickAction(options).Act();
             return this;
         }
 
-        public Scope Hover(Options options = null)
+        internal Click ClickAction(Options options)
         {
-            RetryUntilTimeout(new Hover(this, driver, SetOptions(options)));
+            return new Click(this, driver, SetOptions(options));
+        }
+
+        public virtual ElementScope Hover(Options options = null)
+        {
+            HoverAction(options).Act();
             return this;
         }
 
-        public bool Exists(Options options = null)
+        public virtual ElementScope SendKeys(string keys, Options options = null)
         {
-            return robustWrapper.Robustly(new ElementExistsQuery(this, SetOptions(options)));
+            SendKeysAction(keys, options).Act();
+            return this;
         }
 
-        public bool Missing(Options options = null)
+        internal SendKeys SendKeysAction(string keys, Options options)
         {
-            return robustWrapper.Robustly(new ElementMissingQuery(this, SetOptions(options)));
+            return new SendKeys(keys, this, driver, SetOptions(options));
+        }
+
+        internal Hover HoverAction(Options options)
+        {
+            return new Hover(this, driver, SetOptions(options));
+        }
+
+        public virtual bool Exists(Options options = null)
+        {
+            return ExistsQuery(options).Run();
+        }
+
+        internal ElementExistsQuery ExistsQuery(Options options)
+        {
+            return new ElementExistsQuery(this, SetOptions(options));
+        }
+
+        public virtual bool Missing(Options options = null)
+        {
+            return MissingQuery(options).Run();
+        }
+
+        internal ElementMissingQuery MissingQuery(Options options)
+        {
+            return new ElementMissingQuery(this, SetOptions(options));
         }
     }
 }