+++ src/Coypu/RobustElementScope.cs
@@ -18,5 +18,33 @@ namespace Coypu
         {
             return robustWrapper.Robustly(new ElementQuery(this, options));
         }
+
+        public override ElementScope Click(Options options = null)
+        {
+            RetryUntilTimeout(ClickAction(options));
+            return this;
+        }
+
+        public override ElementScope Hover(Options options = null)
+        {
+            RetryUntilTimeout(HoverAction(options));
+            return this;
+        }
+
+        public override ElementScope SendKeys(string keys, Options options = null)
+        {
+            RetryUntilTimeout(SendKeysAction(keys, options));
+            return this;
+        }
+
+        public override bool Exists(Options options = null)
+        {
+            return robustWrapper.Robustly(ExistsQuery(options));
+        }
+
+        public override bool Missing(Options options = null)
+        {
+            return robustWrapper.Robustly(MissingQuery(options));
+        }
     }
 }